libmetal: amp_demo: linux: update for Versal hardware support
Xilinx Versal hardware's platform manager requires certain nodes to be
requested via a client API before the node is powered on.

One such node which the libmetal_amp_demo demo needs is the TTC0 hardware
block. Update to make sure this block is powered on via client API call
before attempting to use this in demos.

Without this line, the TTC0 block is not powered on and all read
values may be 0.

Signed-off-by: Ben Levinsky <ben.levinsky@xilinx.com>